+++ seatunnel-connectors-v2/connector-elasticsearch/src/main/java/org/apache/seatunnel/connectors/seatunnel/elasticsearch/catalog/ElasticSearchCatalog.java
@@ -150,7 +150,7 @@ public class ElasticSearchCatalog implements Catalog {
                                     fieldName,
                                     elasticSearchDataTypeConvertor.toSeaTunnelType(
                                             fieldName, fieldType),
-                                    null,
+                                    (Long) null,
                                     true,
                                     null,
                                     null);
